+++ src/pages/GroupPage.tsx
@@ -6,7 +6,7 @@ import useBaseStore from "@/stores/baseStore";
 import { KakaoShareButton } from "@/components/KakaoShareBtn";
 import MemberList from "@/components/member/MemberList";
 import { Drawer } from "@/components/ui/drawer";
-import PrayCardDrawer from "@/components/prayCard/PrayCardDrawer";
+import PrayCardContent from "@/components/prayCard/PrayCardContent";
 
 const GroupPage: React.FC = () => {
   const { user } = useAuth();
@@ -58,7 +58,7 @@ const GroupPage: React.FC = () => {
           currentUserId={user?.id}
           groupId={targetGroup?.id}
         ></MemberList>
-        <PrayCardDrawer currentUserId={user!.id} />
+        <PrayCardContent currentUserId={user!.id} />
       </Drawer>
     </div>
   );
